InProgress

+++ TickCrossClient/Pages/GameReqs/GameRequests.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Microsoft.AspNetCore.Mvc.ModelBinding;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -9,6 +10,7 @@ using System.Windows.Data;
 using System.Windows.Documents;
 using System.Windows.Input;
 using System.Windows.Media;
+using System.Windows.Media.Animation;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
@@ -105,7 +107,7 @@ namespace TickCrossClient.Pages.GameReqs
         {
             StackPanel panel = (StackPanel)box.SelectedItem;
 
-            if (panel is null) return null; 
+            if (panel is null) return null;
 
             if (panel is null) return null;
 
@@ -124,6 +126,12 @@ namespace TickCrossClient.Pages.GameReqs
             string toRemoveLogin = GetEnemyLogin(ToRemoveReqsListBox);
             if (toRemoveLogin is null) return;
 
+            if (!await IsGameReqExists(toRemoveLogin, _user.Login))
+            {
+                SetListBoxes();
+                return;
+            }
+
             await ApiService.RemoveGameRequest(_user.Id, toRemoveLogin);
             SetListBoxes();
         }
@@ -136,24 +144,92 @@ namespace TickCrossClient.Pages.GameReqs
         private async void AcceptGameBut_Click(object sender, RoutedEventArgs e)
         {
             //Check is enemy is online or in game!!
-
             string enemyLogin = GetEnemyLogin(SetReqsListBox);
             if (enemyLogin is null) return;
 
             //is user online + is user in game
             bool isEnemyOnline = await ApiService.IsUserIsOnline(enemyLogin);
-            if(!isEnemyOnline)
+            bool isUserOnline = await ApiService.IsUserIsOnline(_user.Login);
+            if (!isEnemyOnline )
             {
-                MessageBox.Show("This user is not online!");
+                MessageBox.Show("Enemy does not have online status!");
                 return;
             }
+            else if (!isUserOnline)
+            {
+                MessageBox.Show("User does not have online status!");
+                return;
+            }
+
+
+            TickCrossLib.Models.User enemy = await ApiService.GetUserByLogin(enemyLogin);
+
+
+            ApiService.SetUserLoginStatus(_user.Id, TickCrossLib.Enums.UserStat.InGame);
+            ApiService.SetUserLoginStatus(enemy.Id, TickCrossLib.Enums.UserStat.InGame);
+    
             SetListBoxes();
+
+            //Start game
+            StartGame(enemyLogin);
+        }
+
+        private async void StartGame(string enemyLogin)
+        {
+            if (!await IsGameReqExists(_user.Login, enemyLogin))
+            {
+                SetListBoxes();
+                return;
+            }
+            //Get game request
+            TickCrossLib.Models.GameRequest req =
+                await ApiService.GetGameRequest(enemyLogin, _user.Login);
+            if (req is null) return;
+
+            //Clear secondary frame
+            ((MainWindow)Window.GetWindow(_frame)).ClearSecondaryFrame();
+
+
+            //Get signs to play
+            List<char>? signs = await ApiService.GetSigns();
+
+            //Create new game
+            TickCrossLib.Models.Game game =
+                new TickCrossLib.Models.Game(req, signs);
+
+            //change status for request
+            await ApiService.SetRequestStatus(req, TickCrossLib.Enums.RequestStatus.Accepted);
+
+            //add game value in DB
+            await ApiService.AddGameInDB(game);
+
+            //Set game id to game value
+            game.AddGameId(await ApiService.GetLastGameId());
+
+            //Create game page
+            GamePage page = new GamePage(game, _frame, _user);
+
+            //Set game page to Main Frame
+            ((MainWindow)Window.GetWindow(_frame)).SetContentToMainFrame(page);
+            ((MainWindow)Window.GetWindow(_frame))._timer.Stop();
+
+            //Start game timer(to make moves
+            page.SetGameMoveTimer(req);
+
+            //Add temp game in DB
+            ApiService.AddTempGameInDB(game.Id);
         }
 
         private async void RejectGameBut_Click(object sender, RoutedEventArgs e)
         {
             string toRemoveLogin = GetEnemyLogin(SetReqsListBox);
             if (toRemoveLogin is null) return;
+
+            if (!await IsGameReqExists(_user.Login, toRemoveLogin))
+            {
+                SetListBoxes();
+                return;
+            }
 
             await ApiService.RejectGameRequest(_user.Id, toRemoveLogin);
             SetListBoxes();
@@ -163,5 +239,12 @@ namespace TickCrossClient.Pages.GameReqs
         {
             SetListBoxes();
         }
+
+        private async Task<bool> IsGameReqExists(string recieverLogin, string senderRequest)
+        {
+            TickCrossLib.Models.GameRequest req = await ApiService.GetGameRequest(senderRequest, recieverLogin);
+            return req is null ? false : true;
+        }
+
     }
 }
